fix(redirects): reorganize redirect configuration and add handling for removed pages
