+++ src/config/redirects/removed-pages.json
@@ -1,0 +1,4 @@
+{
+  "/resources/middleware/connect-rid": "/resources/middleware",
+  "/resources/middleware/csurf": "/resources/middleware"
+}
