--- src/config/removed-pages.json
@@ -1,3 +1,0 @@
-{
-  "/resources/middleware/connect-rid": "/resources/middleware"
-}
